+++ dashboard/models.py
@@ -196,16 +196,11 @@ class DataImport(models.Model):
         ).count()
         self.save()
 
-    def __str__(self) -> str:
-        return (
-            f"Data import #{self.pk} ({localize(localtime(self.start), use_l10n=True)})"
-        )
-
     @property
     def as_dict(self) -> dict[str, Any]:
         return {  # To be consumed on the frontend: we use JS naming conventions
             "id": self.pk,
-            "str": self.__str__(),
+            "name": f"Data import #{self.pk}",
             "startTimestamp": self.start,
         }