+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Current (unreleased)
+
+- Improved "initial data import" filter/selector, according to the suggestions in https://github.com/riparias/gbif-alert/issues/251.
+
 # v1.2.1 (2023-08-21)
 
 - Fixed a display issue with the user menu in the navbar (https://github.com/riparias/gbif-alert/issues/252)